Rename queriesHelper on unit tests to follow snake_case pattern
