+++ framework/test/mocha/integration/dependencies.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const fs = require('fs');
-const QueriesHelper = require('../common/integration/sql/queriesHelper.js');
+const QueriesHelper = require('../common/integration/sql/queries_helper.js');
 const { StorageSandbox } = require('../common/storage_sandbox');
 
 describe('Dependency versions', () => {
