+++ framework/test/mocha/integration/loader/validate_own_chain/full_two_rounds.js
@@ -16,7 +16,7 @@
 
 const Promise = require('bluebird');
 const blockVersion = require('../../../../../src/modules/chain/logic/block_version.js');
-const QueriesHelper = require('../../../common/integration/sql/queriesHelper.js');
+const QueriesHelper = require('../../../common/integration/sql/queries_helper.js');
 const localCommon = require('../../common');
 
 const exceptions = global.exceptions;
