+++ framework/test/mocha/integration/transactions/expire_transactions.js
@@ -18,7 +18,7 @@ const lisk = require('lisk-elements').default;
 const Promise = require('bluebird');
 const randomUtil = require('../../common/utils/random');
 const accountsFixtures = require('../../fixtures/accounts');
-const QueriesHelper = require('../../common/integration/sql/queriesHelper.js');
+const QueriesHelper = require('../../common/integration/sql/queries_helper.js');
 const localCommon = require('../common');
 const Bignum = require('../../../../src/modules/chain/helpers/bignum.js');
